requirements.txt
Cython==0.22.1
Django==1.7
Pillow==2.9.0
argparse==1.2.1
click==4.1
decorator==4.0.2
django-ipware==1.1.0
django-multiupload==0.5
geocode==0.1.0
geocoder==1.5.0
image==1.4.0
numpy==1.9.2
ratelim==0.1.6
requests==2.7.0
six==1.9.0
sorl-thumbnail==12.2
wordcloud==1.1.2
wsgiref==0.1.2
